Added CSV support for loading and saving config data.

+++ src/model/Airport.java
@@ -72,8 +72,8 @@ public final class Airport {
 
     public String getCode() { return code; }
     public String getName() { return name; }
-    public double getX() { return x; }
-    public double getY() { return y; }
+    public int getX() { return x; }
+    public int getY() { return y; }
 
     @Override
     public int hashCode() {
